Change files.size type to bigint
This allow to store files bigger than 4GiB

(cherry picked from commit c56ec56a0de82e799c169d6f14054967b4d79c49)

+++ application/schema/files.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS `%PREFIX%files` (
   `id` int unsigned NOT NULL auto_increment,
   `item_id` int unsigned NOT NULL,
   `order` int(10) unsigned DEFAULT NULL,
-  `size` int unsigned NOT NULL,
+  `size` bigint unsigned NOT NULL,
   `has_derivative_image` tinyint(1) NOT NULL,
   `authentication` char(32) collate utf8_unicode_ci default NULL,
   `mime_type` varchar(255) collate utf8_unicode_ci default NULL,
